dj - rearranged tabs to consolidate InstructorCourseShowPage
