+++ frontend/src/tests/components/TabComponent/EnrollmentTabComponent.test.js
@@ -1,0 +1,445 @@
+import { fireEvent, render, screen, waitFor } from "@testing-library/react";
+import { rosterStudentFixtures } from "fixtures/rosterStudentFixtures";
+import AxiosMockAdapter from "axios-mock-adapter";
+import axios from "axios";
+import {
+  QueryClient,
+  QueryClientProvider,
+  useQuery,
+} from "@tanstack/react-query";
+import EnrollmentTabComponent from "main/components/TabComponent/EnrollmentTabComponent";
+import userEvent from "@testing-library/user-event";
+import { currentUserFixtures } from "fixtures/currentUserFixtures";
+
+const axiosMock = new AxiosMockAdapter(axios);
+const queryClient = new QueryClient();
+const testId = "InstructorCourseShowPage";
+const mockToast = jest.fn();
+jest.mock("react-toastify", () => {
+  const originalModule = jest.requireActual("react-toastify");
+  return {
+    __esModule: true,
+    ...originalModule,
+    toast: (x) => mockToast(x),
+  };
+});
+
+const mockedNavigate = jest.fn();
+jest.mock("react-router", () => ({
+  ...jest.requireActual("react-router"),
+  useNavigate: () => mockedNavigate,
+}));
+
+const ArbitraryTestQueryComponent = () => {
+  const _arbitraryQuery = useQuery({
+    queryKey: ["arbitraryQuery"],
+    queryFn: () => "banana",
+  });
+  return <></>;
+};
+
+describe("EnrollmentTabComponent Tests", () => {
+  beforeEach(() => {
+    axiosMock.reset();
+    axiosMock.resetHistory();
+    queryClient.clear();
+  });
+
+  test("Table Renders", async () => {
+    axiosMock
+      .onGet("/api/rosterstudents/course/1")
+      .reply(200, rosterStudentFixtures.threeStudents);
+
+    render(
+      <QueryClientProvider client={queryClient}>
+        <EnrollmentTabComponent
+          courseId={1}
+          testIdPrefix={testId}
+          currentUser={currentUserFixtures.instructorUser}
+        />
+      </QueryClientProvider>,
+    );
+
+    const rsTestId = "InstructorCourseShowPage-RosterStudentTable";
+
+    await waitFor(() => {
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-0-col-id`),
+      ).toHaveTextContent(rosterStudentFixtures.threeStudents[0].id);
+    });
+
+    const studentFirstName0 = screen.getByText(
+      rosterStudentFixtures.threeStudents[0].firstName,
+    );
+    expect(studentFirstName0).toBeInTheDocument();
+
+    const studentId0 = screen.getByTestId(
+      `${rsTestId}-cell-row-0-col-studentId`,
+    );
+    expect(studentId0).toHaveTextContent(
+      rosterStudentFixtures.threeStudents[0].studentId,
+    );
+  });
+  test("Table Renders with no students", async () => {
+    axiosMock.onGet("/api/rosterstudents/course/7").reply(200, []);
+
+    render(
+      <QueryClientProvider client={queryClient}>
+        <EnrollmentTabComponent
+          courseId={7}
+          testIdPrefix={testId}
+          currentUser={currentUserFixtures.instructorUser}
+        />
+      </QueryClientProvider>,
+    );
+
+    await waitFor(() => {
+      expect(
+        screen.getByTestId(
+          "InstructorCourseShowPage-RosterStudentTable-header-studentId",
+        ),
+      ).toBeInTheDocument();
+    });
+
+    expect(
+      screen.queryByTestId(`${testId}-cell-row-0-col-id`),
+    ).not.toBeInTheDocument();
+
+    const expectedHeaders = ["Student Id", "First Name", "Last Name", "Email"];
+    const expectedFields = ["studentId", "firstName", "lastName", "email"];
+
+    // assert
+    expectedHeaders.forEach((headerText, index) => {
+      const header = screen.getByTestId(
+        `InstructorCourseShowPage-RosterStudentTable-header-${expectedFields[index]}`,
+      );
+      expect(header).toHaveTextContent(headerText);
+    });
+
+    expectedFields.forEach((field) => {
+      const fieldElement = screen.queryByTestId(
+        `${testId}-cell-row-0-col-${field}`,
+      );
+      expect(fieldElement).not.toBeInTheDocument();
+    });
+  });
+
+  test("Successfully makes a call to the backend on submit and clears search filter", async () => {
+    const queryClientSpecific = new QueryClient({
+      defaultOptions: {
+        queries: {
+          retry: false,
+          staleTime: Infinity,
+        },
+      },
+    });
+    const file = new File(["there"], "egrades.csv", { type: "text/csv" });
+
+    axiosMock
+      .onGet("/api/rosterstudents/course/7")
+      .reply(200, rosterStudentFixtures.threeStudents);
+
+    axiosMock.onPost("/api/rosterstudents/upload/egrades").reply(200);
+
+    const user = userEvent.setup();
+    render(
+      <QueryClientProvider client={queryClientSpecific}>
+        <ArbitraryTestQueryComponent />
+        <EnrollmentTabComponent
+          courseId={7}
+          testIdPrefix={testId}
+          currentUser={currentUserFixtures.instructorUser}
+        />
+      </QueryClientProvider>,
+    );
+    await screen.findByTestId("RosterStudentCSVUploadForm-upload");
+
+    const arbitraryUpdateCount = queryClientSpecific.getQueryState([
+      "arbitraryQuery",
+    ]).dataUpdateCount;
+
+    const updateCountStudent = queryClientSpecific.getQueryState([
+      "/api/rosterstudents/course/7",
+    ]).dataUpdateCount;
+
+    // Get the search input and set a search term
+    const searchInput = screen.getByTestId("InstructorCourseShowPage-search");
+    fireEvent.change(searchInput, { target: { value: "test search" } });
+    expect(searchInput.value).toBe("test search");
+
+    const upload = screen.getByTestId("RosterStudentCSVUploadForm-upload");
+    const submitButton = screen.getByTestId(
+      "RosterStudentCSVUploadForm-submit",
+    );
+    await user.upload(upload, file);
+    fireEvent.click(submitButton);
+    await waitFor(() => {
+      expect(axiosMock.history.post[0].params).toEqual({
+        courseId: 7,
+      });
+    });
+    expect(axiosMock.history.post[0].data.get("file")).toEqual(file);
+    expect(mockToast).toBeCalledWith("Roster successfully updated.");
+    expect(
+      queryClientSpecific.getQueryState(["arbitraryQuery"]).dataUpdateCount,
+    ).toBe(arbitraryUpdateCount);
+    expect(
+      queryClientSpecific.getQueryState(["/api/rosterstudents/course/7"])
+        .dataUpdateCount,
+    ).toEqual(updateCountStudent + 1);
+
+    // Verify that the search filter is cleared
+    await waitFor(() => {
+      expect(searchInput.value).toBe("");
+    });
+  });
+
+  test("RosterStudentForm submit works and clears search filter", async () => {
+    const queryClientSpecific = new QueryClient({
+      defaultOptions: {
+        queries: {
+          retry: false,
+          staleTime: Infinity,
+        },
+      },
+    });
+    axiosMock
+      .onGet("/api/rosterstudents/course/7")
+      .reply(200, rosterStudentFixtures.threeStudents);
+
+    axiosMock.onPost("/api/rosterstudents/post").reply(200);
+    render(
+      <QueryClientProvider client={queryClientSpecific}>
+        <ArbitraryTestQueryComponent />
+        <EnrollmentTabComponent
+          courseId={7}
+          testIdPrefix={testId}
+          currentUser={currentUserFixtures.instructorUser}
+        />
+      </QueryClientProvider>,
+    );
+
+    //Great time to check initial values
+    expect(
+      queryClientSpecific.getQueryData(["/api/rosterstudents/course/7"]),
+    ).toStrictEqual([]);
+
+    await screen.findAllByText("Student Id");
+    const arbitraryUpdateCount = queryClientSpecific.getQueryState([
+      "arbitraryQuery",
+    ]).dataUpdateCount;
+    const updateCountStudent = queryClientSpecific.getQueryState([
+      "/api/rosterstudents/course/7",
+    ]).dataUpdateCount;
+
+    // Get the search input and set a search term
+    const searchInput = screen.getByTestId("InstructorCourseShowPage-search");
+    fireEvent.change(searchInput, { target: { value: "test search" } });
+    expect(searchInput.value).toBe("test search");
+
+    expect(screen.queryByText("Cancel")).not.toBeInTheDocument();
+    fireEvent.change(screen.getByLabelText("Student Id"), {
+      target: { value: "123456789" },
+    });
+    fireEvent.change(screen.getByLabelText("First Name"), {
+      target: { value: "Chris" },
+    });
+    fireEvent.change(screen.getByLabelText("Last Name"), {
+      target: { value: "Gaucho" },
+    });
+    fireEvent.change(screen.getByLabelText("Email"), {
+      target: { value: "cgaucho@ucsb.edu" },
+    });
+    fireEvent.click(screen.getByTestId("RosterStudentForm-submit"));
+    await waitFor(() => expect(axiosMock.history.post.length).toEqual(1));
+    expect(axiosMock.history.post[0].params).toEqual({
+      courseId: 7,
+      studentId: "123456789",
+      firstName: "Chris",
+      lastName: "Gaucho",
+      email: "cgaucho@ucsb.edu",
+    });
+    await waitFor(() => expect(mockToast).toBeCalled());
+    expect(mockToast).toBeCalledWith("Roster successfully updated.");
+    expect(
+      queryClientSpecific.getQueryState(["arbitraryQuery"]).dataUpdateCount,
+    ).toBe(arbitraryUpdateCount);
+    expect(
+      queryClientSpecific.getQueryState(["/api/rosterstudents/course/7"])
+        .dataUpdateCount,
+    ).toEqual(updateCountStudent + 1);
+
+    // Verify that the search filter is cleared
+    await waitFor(() => {
+      expect(searchInput.value).toBe("");
+    });
+  });
+
+  describe("Search filter works correctly", () => {
+    const testId = "InstructorCourseShowPage";
+    const rsTestId = "InstructorCourseShowPage-RosterStudentTable";
+    const studentList = [
+      ...rosterStudentFixtures.studentsWithEachStatus,
+      {
+        id: 7,
+        studentId: "A626737",
+        firstName: "Fake",
+        lastName: "Name",
+        email: "fakename@ucsb.edu",
+        githubLogin: "DifferingGitHub",
+        orgStatus: "JOINCOURSE",
+      },
+    ];
+    beforeEach(() => {
+      axiosMock.onGet("/api/rosterstudents/course/1").reply(200, studentList);
+    });
+
+    test("PLaceholder, initial check", async () => {
+      render(
+        <QueryClientProvider client={queryClient}>
+          <EnrollmentTabComponent
+            courseId={1}
+            testIdPrefix={testId}
+            currentUser={currentUserFixtures.instructorUser}
+          />
+        </QueryClientProvider>,
+      );
+      await waitFor(() => {
+        expect(
+          screen.getByTestId(`${rsTestId}-cell-row-0-col-id`),
+        ).toBeInTheDocument();
+      });
+
+      const searchInput = screen.getByTestId(`${testId}-search`);
+      expect(searchInput).toBeInTheDocument();
+      expect(searchInput).toHaveAttribute(
+        "placeholder",
+        "Search by name, email, student ID, or Github Login",
+      );
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-0-col-firstName`),
+      ).toBeInTheDocument();
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-1-col-firstName`),
+      ).toBeInTheDocument();
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-2-col-firstName`),
+      ).toBeInTheDocument();
+    });
+
+    test("First Name, Email", async () => {
+      render(
+        <QueryClientProvider client={queryClient}>
+          <EnrollmentTabComponent
+            courseId={1}
+            testIdPrefix={testId}
+            currentUser={currentUserFixtures.instructorUser}
+          />
+        </QueryClientProvider>,
+      );
+
+      await waitFor(() => {
+        expect(
+          screen.getByTestId(`${rsTestId}-cell-row-0-col-id`),
+        ).toBeInTheDocument();
+      });
+
+      // Verify search input is rendered
+      const searchInput = screen.getByTestId(`${testId}-search`);
+
+      const fullNameStudent = rosterStudentFixtures.studentsWithEachStatus[2]; // Emma Watson
+      fireEvent.change(searchInput, {
+        target: {
+          value:
+            `${fullNameStudent.firstName} ${fullNameStudent.lastName}`.toUpperCase(),
+        },
+      });
+
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-0-col-firstName`),
+      ).toHaveTextContent(fullNameStudent.firstName);
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-0-col-lastName`),
+      ).toHaveTextContent(fullNameStudent.lastName);
+      expect(
+        screen.queryByTestId(`${rsTestId}-cell-row-1-col-firstName`),
+      ).not.toBeInTheDocument();
+
+      fireEvent.change(searchInput, { target: { value: "" } });
+
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-0-col-firstName`),
+      ).toBeInTheDocument();
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-1-col-firstName`),
+      ).toBeInTheDocument();
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-2-col-firstName`),
+      ).toBeInTheDocument();
+
+      fireEvent.change(searchInput, {
+        target: {
+          value:
+            rosterStudentFixtures.studentsWithEachStatus[1].email.toUpperCase(),
+        },
+      });
+
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-0-col-email`),
+      ).toHaveTextContent(
+        rosterStudentFixtures.studentsWithEachStatus[1].email,
+      );
+      expect(
+        screen.queryByTestId(`${rsTestId}-cell-row-1-col-email`),
+      ).not.toBeInTheDocument();
+    });
+
+    test("GitHub Login, Student ID", async () => {
+      render(
+        <QueryClientProvider client={queryClient}>
+          <EnrollmentTabComponent
+            courseId={1}
+            testIdPrefix={testId}
+            currentUser={currentUserFixtures.instructorUser}
+          />
+        </QueryClientProvider>,
+      );
+      await waitFor(() => {
+        expect(
+          screen.getByTestId(`${rsTestId}-cell-row-0-col-id`),
+        ).toBeInTheDocument();
+      });
+
+      const searchInput = screen.getByTestId(`${testId}-search`);
+      const studentWithGithub = studentList[6].githubLogin;
+      fireEvent.change(searchInput, {
+        target: { value: studentWithGithub.toUpperCase() },
+      });
+
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-0-col-githubLogin`),
+      ).toHaveTextContent(studentWithGithub);
+      expect(
+        screen.queryByTestId(`${rsTestId}-cell-row-1-col-firstName`),
+      ).not.toBeInTheDocument();
+
+      fireEvent.change(searchInput, { target: { value: "" } });
+
+      fireEvent.change(searchInput, {
+        target: {
+          value:
+            rosterStudentFixtures.studentsWithEachStatus[1].studentId.toUpperCase(),
+        },
+      });
+
+      expect(
+        screen.getByTestId(`${rsTestId}-cell-row-0-col-studentId`),
+      ).toHaveTextContent(
+        rosterStudentFixtures.studentsWithEachStatus[1].studentId,
+      );
+      expect(
+        screen.queryByTestId(`${rsTestId}-cell-row-1-col-studentId`),
+      ).not.toBeInTheDocument();
+    });
+  });
+});
