+++ frontend/src/tests/pages/Instructor/InstructorCourseShowPage.test.js
@@ -16,7 +16,6 @@ import { systemInfoFixtures } from "fixtures/systemInfoFixtures";
 import axios from "axios";
 import AxiosMockAdapter from "axios-mock-adapter";
 import { rosterStudentFixtures } from "fixtures/rosterStudentFixtures";
-import userEvent from "@testing-library/user-event";
 
 const mockedNavigate = jest.fn();
 jest.mock("react-router", () => ({
@@ -37,8 +36,6 @@ jest.mock("react-toastify", () => {
 });
 
 describe("InstructorCourseShowPage tests", () => {
-  const testId = "InstructorCourseShowPage";
-
   afterEach(() => {
     axiosMock.reset();
     axiosMock.resetHistory();
@@ -63,9 +60,6 @@ describe("InstructorCourseShowPage tests", () => {
       createdByEmail: "phtcon@ucsb.edu",
     };
     axiosMock.onGet("/api/courses/1").reply(200, theCourse);
-    axiosMock
-      .onGet("/api/rosterstudents/course/1")
-      .reply(200, rosterStudentFixtures.threeStudents);
 
     render(
       <QueryClientProvider client={queryClient}>
@@ -95,126 +89,14 @@ describe("InstructorCourseShowPage tests", () => {
       );
     });
 
-    const rsTestId = "InstructorCourseShowPage-RosterStudentTable";
-
-    await waitFor(() => {
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-0-col-id`),
-      ).toHaveTextContent(rosterStudentFixtures.threeStudents[0].id);
-    });
-
     const orgName = screen.getByText("ucsb-cs156-s25");
     expect(orgName).toBeInTheDocument();
 
-    const studentFirstName0 = screen.getByText(
-      rosterStudentFixtures.threeStudents[0].firstName,
-    );
-    expect(studentFirstName0).toBeInTheDocument();
-
-    const studentId0 = screen.getByTestId(
-      `${rsTestId}-cell-row-0-col-studentId`,
-    );
-    expect(studentId0).toHaveTextContent(
-      rosterStudentFixtures.threeStudents[0].studentId,
-    );
     expect(screen.queryByText("Course Not Found")).not.toBeInTheDocument();
 
     jest.advanceTimersByTime(3000);
     expect(mockedNavigate).not.toHaveBeenCalled();
     jest.useRealTimers();
-  });
-
-  test("renders correctly for empty roster student list", async () => {
-    setupInstructorUser();
-
-    axiosMock.onGet("/api/courses/7").reply(200, "");
-
-    axiosMock.onGet("/api/rosterstudents/course/7").reply(200, []);
-
-    render(
-      <QueryClientProvider client={queryClient}>
-        <MemoryRouter initialEntries={["/instructor/courses/7"]}>
-          <Routes>
-            <Route
-              path="/instructor/courses/:id"
-              element={<InstructorCourseShowPage />}
-            />
-          </Routes>
-        </MemoryRouter>
-      </QueryClientProvider>,
-    );
-
-    await waitFor(() => {
-      expect(screen.getByText("Course Name")).toBeInTheDocument();
-    });
-
-    expect(
-      screen.queryByTestId(`${testId}-cell-row-0-col-id`),
-    ).not.toBeInTheDocument();
-
-    const expectedHeaders = ["Student Id", "First Name", "Last Name", "Email"];
-    const expectedFields = ["studentId", "firstName", "lastName", "email"];
-
-    // assert
-    expectedHeaders.forEach((headerText, index) => {
-      const header = screen.getByTestId(
-        `InstructorCourseShowPage-RosterStudentTable-header-${expectedFields[index]}`,
-      );
-      expect(header).toHaveTextContent(headerText);
-    });
-
-    expectedFields.forEach((field) => {
-      const fieldElement = screen.queryByTestId(
-        `${testId}-cell-row-0-col-${field}`,
-      );
-      expect(fieldElement).not.toBeInTheDocument();
-    });
-  });
-
-  test("renders correctly for default empty list when roster student variable is undefined", async () => {
-    setupInstructorUser();
-    axiosMock.onGet("/api/courses/7").reply(200, null);
-
-    axiosMock.onGet("/api/rosterstudents/course/7").reply(200, null);
-
-    render(
-      <QueryClientProvider client={queryClient}>
-        <MemoryRouter initialEntries={["/instructor/courses/7"]}>
-          <Routes>
-            <Route
-              path="/instructor/courses/:id"
-              element={<InstructorCourseShowPage />}
-            />
-          </Routes>
-        </MemoryRouter>
-      </QueryClientProvider>,
-    );
-
-    await waitFor(() => {
-      expect(screen.getByText("Course Name")).toBeInTheDocument();
-    });
-
-    expect(
-      screen.queryByTestId(`${testId}-cell-row-0-col-id`),
-    ).not.toBeInTheDocument();
-
-    const expectedHeaders = ["Student Id", "First Name", "Last Name", "Email"];
-    const expectedFields = ["studentId", "firstName", "lastName", "email"];
-
-    // assert
-    expectedHeaders.forEach((headerText, index) => {
-      const header = screen.getByTestId(
-        `InstructorCourseShowPage-RosterStudentTable-header-${expectedFields[index]}`,
-      );
-      expect(header).toHaveTextContent(headerText);
-    });
-
-    expectedFields.forEach((field) => {
-      const fieldElement = screen.queryByTestId(
-        `${testId}-cell-row-0-col-${field}`,
-      );
-      expect(fieldElement).not.toBeInTheDocument();
-    });
   });
 
   test("handles fallback for courseId", async () => {
@@ -267,9 +149,6 @@ describe("InstructorCourseShowPage tests", () => {
     );
     //Great time to also check initial values
     expect(queryClient.getQueryData(["/api/courses/7"])).toBe(null);
-    expect(queryClient.getQueryData(["/api/rosterstudents/course/7"])).toBe(
-      null,
-    );
 
     await screen.findByText(
       "Course not found. You will be returned to the course list in 3 seconds.",
@@ -322,6 +201,9 @@ describe("InstructorCourseShowPage tests", () => {
     await screen.findByText(
       "Course not found. You will be returned to the course list in 3 seconds.",
     );
+    expect(
+      screen.queryByTestId(`InstructorCourseShowPage-cell-row-0-col-id`),
+    ).not.toBeInTheDocument();
     fireEvent.keyPress(screen.getByText("Course Not Found"), {
       key: "Escape",
       code: 27,
@@ -394,357 +276,7 @@ describe("InstructorCourseShowPage tests", () => {
     fireEvent.click(changeTabs);
   });
 
-  test("Successfully makes a call to the backend on submit and clears search filter", async () => {
-    const queryClientSpecific = new QueryClient({
-      defaultOptions: {
-        queries: {
-          retry: false,
-          staleTime: Infinity,
-        },
-      },
-    });
-    const file = new File(["there"], "egrades.csv", { type: "text/csv" });
-    setupInstructorUser();
-    const theCourse = {
-      ...coursesFixtures.oneCourseWithEachStatus[0],
-      id: 1,
-      createdByEmail: "phtcon@ucsb.edu",
-    };
-
-    axiosMock.onGet("/api/courses/7").reply(200, theCourse);
-
-    axiosMock
-      .onGet("/api/rosterstudents/course/7")
-      .reply(200, rosterStudentFixtures.threeStudents);
-
-    axiosMock.onPost("/api/rosterstudents/upload/egrades").reply(200);
-
-    const user = userEvent.setup();
-    render(
-      <QueryClientProvider client={queryClientSpecific}>
-        <MemoryRouter initialEntries={["/instructor/courses/7"]}>
-          <Routes>
-            <Route
-              path="/instructor/courses/:id"
-              element={<InstructorCourseShowPage />}
-            />
-          </Routes>
-        </MemoryRouter>
-      </QueryClientProvider>,
-    );
-    await screen.findByTestId("RosterStudentCSVUploadForm-upload");
-
-    const alternateUpdateCount = queryClientSpecific.getQueryState([
-      "/api/courses/7",
-    ]).dataUpdateCount;
-    const updateCountStudent = queryClientSpecific.getQueryState([
-      "/api/rosterstudents/course/7",
-    ]).dataUpdateCount;
-
-    // Get the search input and set a search term
-    const searchInput = screen.getByTestId("InstructorCourseShowPage-search");
-    fireEvent.change(searchInput, { target: { value: "test search" } });
-    expect(searchInput.value).toBe("test search");
-
-    const upload = screen.getByTestId("RosterStudentCSVUploadForm-upload");
-    const submitButton = screen.getByTestId(
-      "RosterStudentCSVUploadForm-submit",
-    );
-    await user.upload(upload, file);
-    fireEvent.click(submitButton);
-    await waitFor(() => {
-      expect(axiosMock.history.post[0].params).toEqual({
-        courseId: "7",
-      });
-    });
-    expect(axiosMock.history.post[0].data.get("file")).toEqual(file);
-    expect(mockToast).toBeCalledWith("Roster successfully updated.");
-    expect(
-      queryClientSpecific.getQueryState(["/api/courses/7"]).dataUpdateCount,
-    ).toEqual(alternateUpdateCount);
-    expect(
-      queryClientSpecific.getQueryState(["/api/rosterstudents/course/7"])
-        .dataUpdateCount,
-    ).toEqual(updateCountStudent + 1);
-
-    // Verify that the search filter is cleared
-    await waitFor(() => {
-      expect(searchInput.value).toBe("");
-    });
-  });
-
-  test("RosterStudentForm submit works and clears search filter", async () => {
-    const queryClientSpecific = new QueryClient({
-      defaultOptions: {
-        queries: {
-          retry: false,
-          staleTime: Infinity,
-        },
-      },
-    });
-    setupInstructorUser();
-    const theCourse = {
-      ...coursesFixtures.oneCourseWithEachStatus[0],
-      id: 1,
-      createdByEmail: "phtcon@ucsb.edu",
-    };
-
-    axiosMock.onGet("/api/courses/7").reply(200, theCourse);
-
-    axiosMock
-      .onGet("/api/rosterstudents/course/7")
-      .reply(200, rosterStudentFixtures.threeStudents);
-
-    axiosMock.onPost("/api/rosterstudents/post").reply(200);
-    render(
-      <QueryClientProvider client={queryClientSpecific}>
-        <MemoryRouter initialEntries={["/instructor/courses/7"]}>
-          <Routes>
-            <Route
-              path="/instructor/courses/:id"
-              element={<InstructorCourseShowPage />}
-            />
-          </Routes>
-        </MemoryRouter>
-      </QueryClientProvider>,
-    );
-
-    await screen.findAllByText("Student Id");
-    const alternateUpdateCount = queryClientSpecific.getQueryState([
-      "/api/courses/7",
-    ]).dataUpdateCount;
-    const updateCountStudent = queryClientSpecific.getQueryState([
-      "/api/rosterstudents/course/7",
-    ]).dataUpdateCount;
-
-    // Get the search input and set a search term
-    const searchInput = screen.getByTestId("InstructorCourseShowPage-search");
-    fireEvent.change(searchInput, { target: { value: "test search" } });
-    expect(searchInput.value).toBe("test search");
-
-    expect(screen.queryByText("Cancel")).not.toBeInTheDocument();
-    fireEvent.change(screen.getByLabelText("Student Id"), {
-      target: { value: "123456789" },
-    });
-    fireEvent.change(screen.getByLabelText("First Name"), {
-      target: { value: "Chris" },
-    });
-    fireEvent.change(screen.getByLabelText("Last Name"), {
-      target: { value: "Gaucho" },
-    });
-    fireEvent.change(screen.getByLabelText("Email"), {
-      target: { value: "cgaucho@ucsb.edu" },
-    });
-    fireEvent.click(screen.getByTestId("RosterStudentForm-submit"));
-    await waitFor(() => expect(axiosMock.history.post.length).toEqual(1));
-    expect(axiosMock.history.post[0].params).toEqual({
-      courseId: "7",
-      studentId: "123456789",
-      firstName: "Chris",
-      lastName: "Gaucho",
-      email: "cgaucho@ucsb.edu",
-    });
-    await waitFor(() => expect(mockToast).toBeCalled());
-    expect(mockToast).toBeCalledWith("Roster successfully updated.");
-    expect(
-      queryClientSpecific.getQueryState(["/api/courses/7"]).dataUpdateCount,
-    ).toEqual(alternateUpdateCount);
-    expect(
-      queryClientSpecific.getQueryState(["/api/rosterstudents/course/7"])
-        .dataUpdateCount,
-    ).toEqual(updateCountStudent + 1);
-
-    // Verify that the search filter is cleared
-    await waitFor(() => {
-      expect(searchInput.value).toBe("");
-    });
-  });
-
-  describe("Search filter works correctly", () => {
-    const theCourse = {
-      ...coursesFixtures.oneCourseWithEachStatus[0],
-      id: 1,
-      createdByEmail: "phtcon@ucsb.edu",
-    };
-    const testId = "InstructorCourseShowPage";
-    const rsTestId = "InstructorCourseShowPage-RosterStudentTable";
-    const studentList = [
-      ...rosterStudentFixtures.studentsWithEachStatus,
-      {
-        id: 7,
-        studentId: "A626737",
-        firstName: "Fake",
-        lastName: "Name",
-        email: "fakename@ucsb.edu",
-        githubLogin: "DifferingGitHub",
-        orgStatus: "JOINCOURSE",
-      },
-    ];
-    beforeEach(() => {
-      setupInstructorUser();
-      axiosMock.onGet("/api/courses/1").reply(200, theCourse);
-
-      axiosMock.onGet("/api/rosterstudents/course/1").reply(200, studentList);
-    });
-
-    test("PLaceholder, initial check", async () => {
-      render(
-        <QueryClientProvider client={queryClient}>
-          <MemoryRouter initialEntries={["/instructor/courses/1"]}>
-            <Routes>
-              <Route
-                path="/instructor/courses/:id"
-                element={<InstructorCourseShowPage />}
-              />
-            </Routes>
-          </MemoryRouter>
-        </QueryClientProvider>,
-      );
-      await waitFor(() => {
-        expect(
-          screen.getByTestId(`${rsTestId}-cell-row-0-col-id`),
-        ).toBeInTheDocument();
-      });
-
-      const searchInput = screen.getByTestId(`${testId}-search`);
-      expect(searchInput).toBeInTheDocument();
-      expect(searchInput).toHaveAttribute(
-        "placeholder",
-        "Search by name, email, student ID, or Github Login",
-      );
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-0-col-firstName`),
-      ).toBeInTheDocument();
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-1-col-firstName`),
-      ).toBeInTheDocument();
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-2-col-firstName`),
-      ).toBeInTheDocument();
-    });
-
-    test("First Name, Email", async () => {
-      render(
-        <QueryClientProvider client={queryClient}>
-          <MemoryRouter initialEntries={["/instructor/courses/1"]}>
-            <Routes>
-              <Route
-                path="/instructor/courses/:id"
-                element={<InstructorCourseShowPage />}
-              />
-            </Routes>
-          </MemoryRouter>
-        </QueryClientProvider>,
-      );
-
-      await waitFor(() => {
-        expect(
-          screen.getByTestId(`${rsTestId}-cell-row-0-col-id`),
-        ).toBeInTheDocument();
-      });
-
-      // Verify search input is rendered
-      const searchInput = screen.getByTestId(`${testId}-search`);
-
-      const fullNameStudent = rosterStudentFixtures.studentsWithEachStatus[2]; // Emma Watson
-      fireEvent.change(searchInput, {
-        target: {
-          value:
-            `${fullNameStudent.firstName} ${fullNameStudent.lastName}`.toUpperCase(),
-        },
-      });
-
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-0-col-firstName`),
-      ).toHaveTextContent(fullNameStudent.firstName);
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-0-col-lastName`),
-      ).toHaveTextContent(fullNameStudent.lastName);
-      expect(
-        screen.queryByTestId(`${rsTestId}-cell-row-1-col-firstName`),
-      ).not.toBeInTheDocument();
-
-      fireEvent.change(searchInput, { target: { value: "" } });
-
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-0-col-firstName`),
-      ).toBeInTheDocument();
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-1-col-firstName`),
-      ).toBeInTheDocument();
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-2-col-firstName`),
-      ).toBeInTheDocument();
-
-      fireEvent.change(searchInput, {
-        target: {
-          value:
-            rosterStudentFixtures.studentsWithEachStatus[1].email.toUpperCase(),
-        },
-      });
-
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-0-col-email`),
-      ).toHaveTextContent(
-        rosterStudentFixtures.studentsWithEachStatus[1].email,
-      );
-      expect(
-        screen.queryByTestId(`${rsTestId}-cell-row-1-col-email`),
-      ).not.toBeInTheDocument();
-    });
-
-    test("GitHub Login, Student ID", async () => {
-      render(
-        <QueryClientProvider client={queryClient}>
-          <MemoryRouter initialEntries={["/instructor/courses/1"]}>
-            <Routes>
-              <Route
-                path="/instructor/courses/:id"
-                element={<InstructorCourseShowPage />}
-              />
-            </Routes>
-          </MemoryRouter>
-        </QueryClientProvider>,
-      );
-      await waitFor(() => {
-        expect(
-          screen.getByTestId(`${rsTestId}-cell-row-0-col-id`),
-        ).toBeInTheDocument();
-      });
-
-      const searchInput = screen.getByTestId(`${testId}-search`);
-      const studentWithGithub = studentList[6].githubLogin;
-      fireEvent.change(searchInput, {
-        target: { value: studentWithGithub.toUpperCase() },
-      });
-
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-0-col-githubLogin`),
-      ).toHaveTextContent(studentWithGithub);
-      expect(
-        screen.queryByTestId(`${rsTestId}-cell-row-1-col-firstName`),
-      ).not.toBeInTheDocument();
-
-      fireEvent.change(searchInput, { target: { value: "" } });
-
-      fireEvent.change(searchInput, {
-        target: {
-          value:
-            rosterStudentFixtures.studentsWithEachStatus[1].studentId.toUpperCase(),
-        },
-      });
-
-      expect(
-        screen.getByTestId(`${rsTestId}-cell-row-0-col-studentId`),
-      ).toHaveTextContent(
-        rosterStudentFixtures.studentsWithEachStatus[1].studentId,
-      );
-      expect(
-        screen.queryByTestId(`${rsTestId}-cell-row-1-col-studentId`),
-      ).not.toBeInTheDocument();
-    });
-  });
-  test("Assignment tab is present", async () => {
+  test("Tab Components are Present", async () => {
     const queryClientSpecific = new QueryClient({
       defaultOptions: {
         queries: {
@@ -775,5 +307,8 @@ describe("InstructorCourseShowPage tests", () => {
     );
 
     await screen.findByTestId("AssignmentTabComponent");
+    expect(
+      screen.getByTestId("InstructorCourseShowPage-EnrollmentTabComponent"),
+    ).toBeInTheDocument();
   });
 });
